fix: add explicit float cast

+++ app/src/main/java/com/example/nutritionapp/recommendation/CustomMarkerView.java
@@ -48,7 +48,7 @@ public class CustomMarkerView extends MarkerView {
 
     @Override
     public MPPointF getOffset() {
-        return new MPPointF(-(getWidth() / 2), -getHeight());
+        return new MPPointF(-((float)getWidth() / 2), -getHeight());
     }
 
 }
